pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = "hometask_poetry"
authors = ["Terenteva Viktoria <verba4182@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.lint.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.10.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
